Create Spotify API album image resolver
Updated the code to get all album images for a given category.
It'll now go and get all playlists for a category and then get all album
images for that playlist.

There's some degree of data validation as not all tracks have valid
album data, so we only store what we can.

The result should be a set of all image URLs for a given spotify
category.